Define logic to fill squares
At first, mousedown event allowed for each single
square to be filled per mousedown event which is
not ideal but mouseover only makes the ux annoying

Aiming for a natural user experience along with control,
I customized the event with "isDown" to confirm if
the main button is being press while hovering the
cursor over the squares.

After reading, I found that changing to pointerdown
and pointermove is more cross-device friendly.

This combination has allowed for a seamless experience
of filling the squares.